intro PDRDatasetCacheManager as parent class of PDRCacheManager without   extra threads to ease typing in app configuration classes.
